# Refactor transform