Various clean up and simplification in transform code. As a side-effect, improves performance of `locationPoint` and related methods:

```
# before
locationPoint x 3,028,354 ops/sec ±0.82% (85 runs sampled)
pointLocation x 4,866,038 ops/sec ±1.07% (84 runs sampled)

# after
locationPoint x 4,814,249 ops/sec ±0.72% (89 runs sampled)
pointLocation x 4,535,500 ops/sec ±0.93% (90 runs sampled)
```

👀 @lucaswoj