@mourner 2016-11-30
This is just the first step — the next one will be fixing #2868 by refactoring the overscaling code.