@lucaswoj 2016-12-01
LGTM other than the concern about de-parameterizing `worldSize` ☝️