@anandthakker 2017-03-09
@mourner you mentioned [here](https://github.com/mapbox/mapbox-gl-js/pull/3721#issuecomment-263976227) something about refactoring the overscaling code (and thus resolving this issue).  Do you have a plan/approach in mind for that?