@mourner 2017-03-09
@anandthakker I attempted it a few weeks ago but got lost and would need to start again. The plan is to never treat the `z` coordinate of a tile object differently depending on context, and pass a separate zoom value (from the current view) for evaluating style and overscaling tiles.